+++ src/pages/Home/Home.css
@@ -1,0 +1,127 @@
+
+ /* HOME PAGE */
+
+ .home-page {
+  height: 100vh;
+  width: 100vw;
+  position: absolute;
+  display: block;
+ }
+
+ .home-page-background-1 {
+  padding-top: 180px;
+  padding-bottom: 60px;
+  height: fit-content;
+  width: 100vw;
+  background-image: url(../../@types/assets/hero-blob-scene.svg);
+  background-size: cover;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+ }
+
+ .home-page-container-1 {
+  display: flex;
+  position: relative;
+  align-items: center;
+  justify-content: center;
+  width: 65vw;
+  max-width: 750px;
+ }
+
+ .home-page-title {
+  padding-bottom: 30px;
+  color: white;
+  position: relative;
+  display: block;
+  justify-content: center;
+ }
+
+ .home-page-subtitle {
+  color: white;
+  padding-bottom: 25px;
+  width: 90%;
+  position: relative;
+  display: block;
+  justify-content: center;
+  align-self: center;
+ }
+
+ .home-page-subtitle h3 {
+  color: var(--lightgrey);
+ }
+
+ .button-box {
+  display: flex;
+  position: relative;
+  justify-content: center;
+  align-self: center;
+ }
+
+ .button-col {
+  display: flex;
+  justify-content: center;
+ }
+
+
+ .home-page-background-2 {
+  height: fit-content;
+  width: 100vw;
+  background-color: var(--charcoal);
+  background-size: cover;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+ }
+
+ .home-page-container-2 {
+  display: block;
+  position: relative;
+  align-items: center;
+  padding-bottom: 100px;
+  padding-top: 80px;
+  color: var(--lightgrey);
+  width: 65vw;
+  text-align: left;
+ }
+
+ .home-page-col {
+  width: 30vw;
+ }
+
+ .home-page-container-3 {
+  display: block;
+  position: relative;
+  align-items: center;
+  padding-bottom: 60px;
+  color: var(--lightgrey);
+  width: 60vw;
+ }
+
+ .faq-heading {
+  padding-bottom: 20px;
+ }
+
+ .home-page-background-3 {
+  padding-top: 60px;
+  height: fit-content;
+  width: 100vw;
+  background-image: url(../../../src/@types/assets/cta-blob-scene.svg);
+  background-size: cover;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+ }
+
+ .home-page-container-4 {
+  display: block;
+  position: relative;
+  align-items: center;
+  padding-bottom: 30px;
+  color: var(--lightgrey);
+  width: 65vw;
+  max-width: 750px;
+ }
