+++ src/pages/NoRoute/NoRoute.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import './NoRoute.css';
 
 function NoRoute() {
     return (
