+++ src/pages/TaskPage/TaskPage.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import "./TaskPage.css";
 
 function TaskPage() {
     return (
